Refactor queue configuration, allow options to be passed
Add the ability to pass in arbitrary queue-specific options.

I wanted the ability to give Bull-specific options to Arena. For my
specific use case, it was to pass in the createClient option in order to
reduce the number of open Redis connections in my application. This adds
that ability, and also breaks apart the Queues class into two tested
classes.

For testing, I used ava-spec, sinon, and sandboxed-module because
they're used by bee-queue.