Fix using stale Redis stats, parse info directly
The ioredis client used by bull makes no guarantees that the
`serverInfo` field be updated upon issuing an INFO command, and in fact
does not appear to do so (given bull 3.13.0 and ioredis 4.16.3).

While the `serverInfo` field is indeed populated after the initial
connection is established, it is not updated by calling INFO as assumed
with the previous logic. This results in stale data being presented by
Arena.

Rather than relying on this undocumented and evidently fragile
behaviour, we can instead parse the result of the INFO command on our
own.

This also happens to step around an issue with the way that ioredis
parses key-value data in handling the INFO response, where values
containing the colon character are incorrectly trimmed.

Given the stability and superb backwards-compatibility of Redis, this
should be a highly reliable solution.